+++ changelog.md
@@ -75,7 +75,7 @@
 
 - Added `euclDiv` and `euclMod` to `math`.
 - Added `httpcore.is1xx` and missing HTTP codes.
-- Added `jsconsole.doAssert` for JavaScript target.
+- Added `jsconsole.jsAssert` for JavaScript target.
 
 
 ## Language changes